phu/api-users

+++ database/migrations/2023_06_02_140256_create_orders_details.php
@@ -14,8 +14,8 @@ return new class extends Migration
         Schema::create('orders_details', function (Blueprint $table) {
             $table->id();
 
-            $table->integer('order_id')->unsigned();
-            $table->integer('product_id')->unsigned();
+            $table->unsignedBigInteger('order_id')->unsigned();
+            $table->unsignedBigInteger('product_id')->unsigned();
             $table->integer('qty');
             $table->double('total');
             $table->tinyText('status');
